finished part 3

+++ server/controllers/videoController.js
@@ -5,15 +5,13 @@ const Util = require("../util/util.js");
 
 async function getFeed(req, res) {
   try {
-    console.log("Fetching...");
+    console.log("Fetching feed...");
     const numberOfVideos = 10;
-    console.log("Top 10 started");
     const mostViewed = await VideoService.getTopVideos(numberOfVideos);
-    console.log("10 other started");
     const unchosenVideos = await VideoService.getUnchosenVideos(numberOfVideos, mostViewed);
     const videoList = Util.randomizeArray([...mostViewed, ...unchosenVideos]);
 
-    console.log("Fetching video list ended successfully");
+    console.log("Fetching videos ended successfully");
     res.status(200).json(videoList);
   } catch (error) {
     console.error("Error fetching video list: ", error);
@@ -26,8 +24,9 @@ async function getFeed(req, res) {
 async function getUserVideoList(req, res) {
   const userId = req.params.id;
   try {
+    console.log("Fetching user video list...");
     const userVideoList = await VideoService.getUserVideoList(userId);
-    console.log("Fetched creator video list successfully");
+    console.log("Fetching user video list successfully");
     res.status(200).json(userVideoList);
   } catch (error) {
     console.error("Error fetching creator video list:", userId, error);
@@ -57,6 +56,13 @@ async function getVideo(req, res) {
 }
 
 async function updateVideo(req, res) {
+  const userId = req.params.id;
+  const authUser = req.user;
+
+  if (authUser._id.toString() !== userId) {
+    return res.status(401).json({ message: "Authentication required to update video" });
+  }
+
   const videoId = req.params.pid;
   const newData = req.body;
 
@@ -76,7 +82,15 @@ async function updateVideo(req, res) {
 }
 
 async function createVideo(req, res) {
+  const userId = req.params.id;
+  const authUser = req.user;
+
+  if (authUser._id.toString() !== userId) {
+    return res.status(401).json({ message: "Authentication required to create a video" });
+  }
+
   const video = req.body;
+
   try {
     const newVideo = await Video.createVideo(video);
     const videoData = await Video.findById(newVideo._id).select("-comments -video");
@@ -92,12 +106,17 @@ async function createVideo(req, res) {
 }
 
 async function deleteVideo(req, res) {
-  const videoId = req.params.pid;
   const userId = req.params.id;
+  const authUser = req.user;
+
+  if (authUser._id.toString() !== userId) {
+    return res.status(401).json({ message: "Authentication required to delete video" });
+  }
+
+  const videoId = req.params.pid;
 
   try {
     await Video.deleteVideo(videoId, userId);
-    console.log("Deleted video successfully");
     res.status(200).json({ message: "Video deleted successfully" });
   } catch (error) {
     console.error("Error deleting video:", videoId, error);
@@ -126,7 +145,6 @@ async function deleteAllVideos(userId) {
 }
 
 async function filterVideos(req, res) {
-  const videoId = req.params.pid;
   const { search, text } = req.body;
 
   try {
@@ -151,7 +169,7 @@ async function filterVideos(req, res) {
         }
       ).lean();
     }
-
+    console.log(filtered);
     res.status(200).json(filtered);
   } catch (error) {
     console.error("Error filtering videos");
